# Conflicts:
#	packages/shared-components/src/components/attachment/AttachmentUpload.tsx
#	packages/shared-components/src/pages/FyllUtRouter.tsx
#	packages/shared-components/src/pages/upload-personal-id/UploadPersonalIdPage.tsx
#	packages/shared-domain/src/file/index.ts

+++ packages/shared-components/src/components/attachment/AttachmentUploadContext.tsx
@@ -1,14 +1,15 @@
 import { FileObject } from '@navikt/ds-react';
 import { Submission, UploadedFile } from '@navikt/skjemadigitalisering-shared-domain';
 import { createContext, useContext, useState } from 'react';
-import { deleteFile, uploadFile } from '../../api/nologin-file-upload/nologinFileUpload';
+import { deleteAttachment, deleteFile, uploadFile } from '../../api/nologin-file-upload/nologinFileUpload';
 import { useForm } from '../../context/form/FormContext';
 import { useSendInn } from '../../context/sendInn/sendInnContext';
 
 interface AttachmentUploadContextType {
-  handleUploadFiles: (vedleggId: string, files: FileObject[]) => Promise<void>;
-  handleDeleteFile: (vedleggId: string, fileId: string) => Promise<void>;
-  addError: (vedleggId: string, error: string) => void;
+  handleUploadFiles: (attachmentId: string, files: FileObject[]) => Promise<void>;
+  handleDeleteFile: (attachmentId: string, fileId: string) => Promise<void>;
+  handleDeleteAttachment: (attachmentId: string) => Promise<void>;
+  addError: (attachmentId: string, error: string) => void;
   uploadedFiles: UploadedFile[];
   errors: Record<string, string | undefined>;
 }
@@ -16,6 +17,7 @@ interface AttachmentUploadContextType {
 const initialContext: AttachmentUploadContextType = {
   handleUploadFiles: async () => {},
   handleDeleteFile: async () => {},
+  handleDeleteAttachment: async () => {},
   addError: () => {},
   uploadedFiles: [],
   errors: {},
@@ -34,60 +36,87 @@ const AttachmentUploadProvider = ({ children }: { children: React.ReactNode }) =
       (current) => ({ ...current, uploadedFiles: [...(current?.uploadedFiles ?? []), ...files] }) as Submission,
     );
   };
-  const deleteFromSubmission = (fileId: string) => {
+  const removeFileFromSubmission = (fileId: string) => {
     setSubmission(
       (current) =>
         ({
           ...current,
-          uploadedFiles: (current?.uploadedFiles ?? []).filter((file) => file.filId !== fileId),
+          uploadedFiles: (current?.uploadedFiles ?? []).filter((file) => file.fileId !== fileId),
+        }) as Submission,
+    );
+  };
+
+  const removeFilesFromSubmission = (attachmentId: string) => {
+    setSubmission(
+      (current) =>
+        ({
+          ...current,
+          uploadedFiles: (current?.uploadedFiles ?? []).filter((file) => file.attachmentId !== attachmentId),
         }) as Submission,
     );
   };
 
   console.log(uploadedFiles);
-  const addError = (vedleggId: string, error: string) => {
+  const addError = (attachmentId: string, error: string) => {
     setErrors((prev) => ({
       ...prev,
-      [vedleggId]: error,
+      [attachmentId]: error,
     }));
   };
 
-  const removeError = (vedleggId: string) => {
+  const removeError = (attachmentId: string) => {
     setErrors((prev) => {
-      const { [vedleggId]: _, ...rest } = prev; // Remove the error for the specific vedleggId
+      const { [attachmentId]: _, ...rest } = prev; // Remove the error for the specific attachmentId
       return rest;
     });
   };
 
-  const handleUploadFiles = async (vedleggId: string, files: FileObject[]) => {
+  const handleUploadFiles = async (attachmentId: string, files: FileObject[]) => {
     const result = await Promise.all(
       files.map(async ({ file }) => {
         try {
-          return await uploadFile(file, vedleggId);
+          return await uploadFile(file, attachmentId);
         } catch (_e) {
-          addError(vedleggId, 'Det oppstod en feil under opplasting av filen. Prøv igjen senere.');
+          addError(attachmentId, 'Det oppstod en feil under opplasting av filen. Prøv igjen senere.');
         }
       }),
     );
     const successfulUploads = result.filter((file): file is UploadedFile => file !== undefined);
     if (successfulUploads.length === result.length) {
-      removeError(vedleggId);
+      removeError(attachmentId);
     }
     addToSubmission(successfulUploads);
     setInnsendingsId(successfulUploads[0]?.innsendingId);
   };
 
-  const handleDeleteFile = async (vedleggId: string, fileId: string) => {
+  const handleDeleteFile = async (attachmentId: string, fileId: string) => {
     try {
+      if (!innsendingsId) {
+        throw new Error('InnsendingsId is not set');
+      }
       await deleteFile(fileId, innsendingsId);
-      removeError(vedleggId);
-      deleteFromSubmission(fileId);
+      removeError(attachmentId);
+      removeFileFromSubmission(fileId);
     } catch (_e) {
-      addError(vedleggId, 'Det oppstod en feil under sletting av filen. Prøv igjen senere.');
+      addError(attachmentId, 'Det oppstod en feil under sletting av filen. Prøv igjen senere.');
     }
   };
 
-  const value = { handleUploadFiles, handleDeleteFile, addError, uploadedFiles, errors };
+  const handleDeleteAttachment = async (attachmentId: string) => {
+    try {
+      if (!innsendingsId) {
+        throw new Error('InnsendingsId is not set');
+      }
+      await deleteAttachment(attachmentId, innsendingsId);
+      removeError(attachmentId);
+      removeFilesFromSubmission(attachmentId);
+    } catch (e) {
+      addError(attachmentId, 'Det oppstod en feil under sletting av vedlegget. Prøv igjen senere.');
+      throw e;
+    }
+  };
+
+  const value = { handleUploadFiles, handleDeleteFile, handleDeleteAttachment, addError, uploadedFiles, errors };
   return <AttachmentUploadContext.Provider value={value}>{children}</AttachmentUploadContext.Provider>;
 };
